requirements.txt
aiofiles==0.6.0
click==7.1.2
dnspython==2.1.0
ecdsa==0.14.1
email-validator==1.1.2
fastapi==0.63.0
h11==0.12.0
idna==3.1
Jinja2==2.11.3
MarkupSafe==1.1.1
pyasn1==0.4.8
pydantic==1.8.1
python-decouple==3.4
python-jose==3.2.0
rsa==4.7.2
six==1.15.0
starlette==0.13.6
typing-extensions==3.7.4.3
uvicorn==0.13.4
